refactor(table): 调整 TS 类型定义位置
