+++ src/table/hooks/useFixed.ts
@@ -6,47 +6,7 @@ import { BaseTableCol, TdBaseTableProps } from '../type';
 import getScrollbarWidth from '../../_common/js/utils/getScrollbarWidth';
 import { on, off } from '../../utils/dom';
 import { TDisplayNoneElementRefresh } from '../../hooks/useDestroyOnClose';
-
-export interface ColumnStickyLeftAndRight {
-  left: number[];
-  right: number[];
-  top: number[];
-  bottom?: number[];
-}
-
-export interface TableColFixedClasses {
-  left: string;
-  right: string;
-  lastLeft: string;
-  firstRight: string;
-  leftShadow: string;
-  rightShadow: string;
-}
-
-export interface TableRowFixedClasses {
-  top: string;
-  bottom: string;
-  firstBottom: string;
-  withoutBorderBottom: string;
-}
-
-export interface FixedColumnInfo {
-  left?: number;
-  right?: number;
-  top?: number;
-  bottom?: number;
-  parent?: FixedColumnInfo;
-  children?: string[];
-  width?: number;
-  height?: number;
-  col?: BaseTableCol;
-  index?: number;
-  lastLeftFixedCol?: boolean;
-  firstRightFixedCol?: boolean;
-}
-
-// 固定表头和固定列 具体的固定位置（left/top/right/bottom）
-export type RowAndColFixedPosition = Map<string | number, FixedColumnInfo>;
+import { TableColFixedClasses, TableRowFixedClasses, FixedColumnInfo, RowAndColFixedPosition } from '../interface';
 
 // 固定列相关类名处理
 export function getColumnFixedStyles(
