+++ src/table/tbody.tsx
@@ -5,8 +5,7 @@ import pick from 'lodash/pick';
 import TrElement, { ROW_LISTENERS, TABLE_PROPS } from './tr';
 import { useConfig } from '../hooks/useConfig';
 import { RowspanColspan, TableRowData, BaseTableCellParams } from './type';
-import { BaseTableProps } from './interface';
-import { RowAndColFixedPosition } from './hooks/useFixed';
+import { BaseTableProps, RowAndColFixedPosition } from './interface';
 import { useTNodeJSX } from '../hooks/tnode';
 import useClassName from './hooks/useClassName';
 import baseTableProps from './base-table-props';
